fixed U2 A1, 4, and the test

+++ src/DataSet.java
@@ -39,6 +39,6 @@ public class DataSet
 
 	public double average()
 	{
-		return ((double) sum / (double) count);
+		return (double) sum / count;
 	}
 }
